pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gaussbio3d"
version = "0.1.0"
description = "Multiscale Gauss Linking Integral Library for Biomolecular 3D Topology"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
dependencies = [
  "numpy>=1.20.0",
  "biopython>=1.79",
]

[project.optional-dependencies]
chem = [
  "rdkit-pypi>=2021.9.1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["examples*", "tests*", "docs*"]